Add Simulation task for the main wrapper function.
Up until now we were using the main wrapper only to fit a model. I have updated the code to not only fit but allow for two more tasks: simualtion and evaluation. The evaluation task simply retrun the likelihood evaluated at the given point.

The simulation task runs a simulation for a given parameter scheme using parallel computation. This may be an issue for MAC users.