Rewrite the PArtcileFilter_Res_Ar function
This function essentially implements our algorithm. It was written a lot like spaggetii code, with many changes upon. I rewrote it using some individual functions for the steps of the algorithm. The way the funcityon is written right now one can read the code and easily match it to the steps of the algorithm.