Implemented the calculation of the predictive distribution for Issue #25 and added Dominick data.
The predictive dist is in the function PDValues see relation (36). I added one test for this, but I need to test it more thoroughly. I may need to think about the values ofr the first p+q timem points. Should they be zero?

I added the Dominick Sales data as RData and adjusted the tests to read this instead of the csv.