Mixed Poisson covariate and initial estimate
This is related to #16, #26 and #10.

For the covariates I allow lambda1 and lambda2 to change in time based on the same covariate x_t, while the mixing prob is constant.

For the initial estimates I used the package mixtools that has an EME algorithm, but also verified some results with flexmix package (commented out).

I also added some tests comparing my pmf, cdf, distributions with the the ones in extraDistr package. The extraDistr package, however, doesn't have an inverse cdf, I had to write one that allowed for vector inputs.