Working on #65 and #84
Adding an example for FitMultiplePF_Res_New function.

I did some work to allow for multiple PartcileFilter values. This wa initially possible, and in fact the purpose of the FitMultiplePFRes functin was exactly that, for someone to be able to fit the log likeliuhood multiple times (for diferent PArticle number s AND for a requestednnumber of fits). When I create the PrepareOutput funciton I ghuess I didnt preserve this functionality and it wasnt working properly.